OCR API 2/requirements.txt
fastapi 
uvicorn
python-multipart
pdf2image
packaging
shapely
requests
websockets
httpx
docstring-parser
pyproject_hooks
wrapt 
asgiref
python-socketio[asgi]
fastapi-socketio
google-cloud-vision